abstracted useGiteaReactToolkit.js from AppContext
